requirements.txt
python-telegram-bot
tensorflow